Refactors TS
