+++ models/experimental/SSR/tt/__init__.py
@@ -1,10 +1,12 @@
-from .mlp import TTMlp
-from .window_attn import TTWindowAttention
-from .swin_transformer_block import TTSwinTransformerBlock
-from .patch_embed import TTPatchEmbed
-from .patch_merging import TTPatchMerging
-from .basic_block import TTBasicLayer
-from .mask_token_inference import TTMaskTokenInference
+from .common import TTMlp
+from .tile_selection import (
+    TTWindowAttention,
+    TTSwinTransformerBlock,
+    TTPatchEmbed,
+    TTPatchMerging,
+    TTBasicLayer,
+    TTMaskTokenInference,
+)
 from .patch_unembed import TTPatchUnEmbed
 from .window_attn_tr import TTWindowAttentionTR
