+++ models/experimental/SSR/tt/tile_selection/basic_layer/swin_transformer_block/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+# SPDX-License-Identifier: Apache-2.0
+
+from .swin_transformer_block import TTSwinTransformerBlock
+from .window_attn import TTWindowAttention
+
+__all__ = ["TTSwinTransformerBlock", "TTWindowAttention"]
